450: Add horseshoe prior, as well as constant prior (#472)
* fix math printing in design doc

* add stats spec

* add prior_horseshoe - first stab

* explain syntax test part

* add a test for it

* add shrinkage factors as transformed parameters

* make sure all shrinkage factors in samples

* add covariates and shrinkage gen. functions and methods, run docs

* first stab at comparison with brms but does not run yet

* fix model.matrix for models without intercept

* set lambda effectively to 1 in jmpost comparison with brms

* truncation, functions

* move shrinkage factor calc to generated quantities to save sampling time

* ensure to return empty stan module for empty prior to avoid warning

* Add prior const (#470)

* slight tangent: add prior_const for surv exp first as a try

* first matching comparison between jmpost and brms for horseshoe prior usage

* first try to reduce to one model.stan file

* let Prior handle parameter declaration in the Stan code together with limits

* adapt documentation

* add snapshot tests for model stan code construction

* add vignette

* few fixes

* some test fixes

* some fixes

* fix prior_const printing

* prior-aware dispatch for render_stan_limits(), so T[0, ] is omitted for:
prior_cauchy(), prior_gamma(), prior_invgamma(), prior_loglogistic(), prior_lognormal()

* updates to tests

* fix one test

* use positive epsilon for the one test

* make test less rigid

* polish vignette

* add initial values also for auxiliary prior parameters (needed for horseshoe prior) to avoid warnings about uninitialized parameters

* polish vignette

* fix linters

* fix docs and spelling

* fix docs

* fix snaps

* add missing news entry for horseshoe prior

+++ R/DataSurvival.R
@@ -128,9 +128,24 @@ model.matrix.DataSurvival <- function(
 ) {
     design_mat <- stats::model.matrix(vars$frm, data = df)
     remove_index <- grep("(Intercept)", colnames(design_mat), fixed = TRUE)
-    design_mat <- design_mat[, -remove_index, drop = FALSE]
+    if (length(remove_index) > 0) {
+        design_mat <- design_mat[, -remove_index, drop = FALSE]
+    }
     rownames(design_mat) <- NULL
     design_mat
+}
+
+#' Extract Covariate Names for Survival Data
+#'
+#' @typed object: DataSurvival
+#'  Survival data object to extract covariate names from.
+#' @param ... not used.
+#'
+#' @family covariates
+#' @export
+covariates.DataSurvival <- function(object, ...) {
+    design_mat <- model.matrix(object, ...)
+    colnames(design_mat)
 }
 
 #' @rdname as_stan_list.DataObject
